Add native omnibar controller and layout coordinator
